// src/generated/prisma/models/Review.ts
export type ReviewWhereInput = {
  AND?: Prisma.ReviewWhereInput | Prisma.ReviewWhereInput[]
  OR?: Prisma.ReviewWhereInput[]
  NOT?: Prisma.ReviewWhereInput | Prisma.ReviewWhereInput[]
  id?: Prisma.StringFilter<"Review"> | string
  productId?: Prisma.StringFilter<"Review"> | string
  userId?: Prisma.StringFilter<"Review"> | string
  rating?: Prisma.IntFilter<"Review"> | number
  title?: Prisma.StringNullableFilter<"Review"> | string | null
  body?: Prisma.StringNullableFilter<"Review"> | string | null
  status?: Prisma.EnumReviewStatusFilter<"Review"> | $Enums.ReviewStatus
  isVerifiedPurchase?: Prisma.BoolFilter<"Review"> | boolean
  createdAt?: Prisma.DateTimeFilter<"Review"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Review"> | Date | string
  product?: Prisma.XOR<Prisma.ProductScalarRelationFilter, Prisma.ProductWhereInput>
  user?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
}

export type ReviewOrderByWithRelationInput = {
  id?: Prisma.SortOrder
  productId?: Prisma.SortOrder
  userId?: Prisma.SortOrder
  rating?: Prisma.SortOrder
  title?: Prisma.SortOrderInput | Prisma.SortOrder
  body?: Prisma.SortOrderInput | Prisma.SortOrder
  status?: Prisma.SortOrder
  isVerifiedPurchase?: Prisma.SortOrder
  createdAt?: Prisma.SortOrder
  updatedAt?: Prisma.SortOrder
  product?: Prisma.ProductOrderByWithRelationInput
  user?: Prisma.UserOrderByWithRelationInput
}